feat(tokens): handle specific token symbols and addresses in useToken function
This commit modifies the useToken function in the Tokens.ts file to handle specific token symbols and addresses. It also adds the ChainId import and the USDC_E and USDT_E constants. Changes are also made to the AddLiquidity index file, removing the "showCommonBases" attribute from two input fields. The friendly name and token name for the AUUSDC_AUUSDT pool in the constants file are updated. In the stake-constants.ts file, USDC_E and USDT_E tokens are added and replace USDC and USDT tokens in various parts of the file.